s/ReedSolomon/CheckedParams/
